+++ aiapy/calibrate/__init__.py
@@ -2,8 +2,8 @@
 Subpackage for calibrating AIA imaging data.
 """
 
-from .meta import *  # NOQA
-from .prep import *  # NOQA
-from .spikes import *  # NOQA
-from .transform import *  # NOQA
-from .uncertainty import *  # NOQA
+from .meta import *  # NOQA: F403
+from .prep import *  # NOQA: F403
+from .spikes import *  # NOQA: F403
+from .transform import *  # NOQA: F403
+from .uncertainty import *  # NOQA: F403