+++ aiapy/calibrate/meta.py
@@ -56,7 +56,7 @@ def fix_observer_location(smap):
     new_meta["hglt_obs"] = coord.lat.to(u.degree).value
     new_meta["dsun_obs"] = coord.radius.to(u.m).value
 
-    return smap._new_instance(smap.data, new_meta, plot_settings=smap.plot_settings, mask=smap.mask)
+    return smap._new_instance(smap.data, new_meta, plot_settings=smap.plot_settings, mask=smap.mask)  # NOQA: SLF001
 
 
 def update_pointing(smap, *, pointing_table=None):
@@ -104,10 +104,12 @@ def update_pointing(smap, *, pointing_table=None):
     `aiapy.calibrate.util.get_pointing_table`
     """
     if not contains_full_disk(smap):
-        raise ValueError("Input must be a full disk image.")
+        msg = "Input must be a full disk image."
+        raise ValueError(msg)
     shape_full_frame = (4096, 4096)
-    if not all(d == (s * u.pixel) for d, s in zip(smap.dimensions, shape_full_frame)):
-        raise ValueError(f"Input must be at the full resolution of {shape_full_frame}")
+    if not all(d == (s * u.pixel) for d, s in zip(smap.dimensions, shape_full_frame, strict=True)):
+        msg = f"Input must be at the full resolution of {shape_full_frame}"
+        raise ValueError(msg)
     if pointing_table is None:
         # Make range wide enough to get closest 3-hour pointing
         pointing_table = get_pointing_table(smap.date - 12 * u.h, smap.date + 12 * u.h)
@@ -134,10 +136,13 @@ def update_pointing(smap, *, pointing_table=None):
     t_obs = astropy.time.Time(t_obs)
     t_obs_in_interval = np.logical_and(t_obs >= pointing_table["T_START"], t_obs < pointing_table["T_STOP"])
     if not t_obs_in_interval.any():
-        raise IndexError(
+        msg = (
             f"No valid entries for {t_obs} in pointing table "
             f'with first T_START date of {pointing_table[0]["T_START"]} '
-            f'and a last T_STOP date of {pointing_table[-1]["T_STOP"]}.',
+            f'and a last T_STOP date of {pointing_table[-1]["T_STOP"]}.'
+        )
+        raise IndexError(
+            msg,
         )
     i_nearest = np.where(t_obs_in_interval)[0][0]
     w_str = f"{smap.wavelength.to(u.angstrom).value:03.0f}"
@@ -185,4 +190,4 @@ def update_pointing(smap, *, pointing_table=None):
     new_meta.pop("PC1_2")
     new_meta.pop("PC2_1")
     new_meta.pop("PC2_2")
-    return smap._new_instance(smap.data, new_meta, plot_settings=smap.plot_settings, mask=smap.mask)
+    return smap._new_instance(smap.data, new_meta, plot_settings=smap.plot_settings, mask=smap.mask)  # NOQA: SLF001