+++ aiapy/data/sample.py
@@ -1,34 +1,79 @@
 """
-This module provides the following sample data files. These files are
-downloaded when this module is imported for the first time.
+This module provides the following sample data files.
 
+When a sample shortname is accessed, the corresponding file is downloaded if needed.
+All files can be downloaded by calling :func:`~irispy.data.sample.download_all`.
+
+Summary variables
+-----------------
+.. list-table::
+   :widths: auto
+
+   * - ``file_dict``
+     - Dictionary of all sample shortnames and, if downloaded, corresponding
+       file locations on disk (otherwise, ``None``)
+   * - ``file_list``
+     - List of disk locations for sample data files that have been downloaded
+
+Sample shortnames
+-----------------
 .. list-table::
    :widths: auto
    :header-rows: 1
 
-   * - Variable name
+   * - Sample shortname
      - Name of downloaded file
 """
 
-import sys
-from pathlib import Path
+from ._sample import _SAMPLE_DATA, _get_sample_files
 
-from ._sample import _SAMPLE_FILES, download_sample_data
+# Add a table row to the module docstring for each sample file
+for _keyname, _filename in sorted(_SAMPLE_DATA.items()):
+    __doc__ += f"   * - ``{_keyname}``\n     - {_filename}\n"  # NOQA: A001
 
-files = download_sample_data()
-file_dict = {}
-for f in files:
-    name = Path(f).name
-    _key = _SAMPLE_FILES.get(name, None)
-    if _key:
-        setattr(sys.modules[__name__], _key, str(f))
-        file_dict.update({_key: f})
 
-# Sort the entries in the dictionary
-file_dict = dict(sorted(file_dict.items()))
-file_list = file_dict.values()
+# file_dict and file_list are not normal variables; see __getattr__() below
+__all__ = [  # NOQA: PLE0604, F822
+    "download_all",
+    "file_dict",
+    "file_list",
+    *sorted(_SAMPLE_DATA.keys()),
+]
 
-for keyname, filename in file_dict.items():
-    __doc__ += f"   * - ``{keyname}``\n     - {Path(filename).name}\n"
 
-__all__ = [*list(_SAMPLE_FILES.values()), "file_dict", "file_list"]
+# See PEP 562 (https://peps.python.org/pep-0562/) for module-level __dir__()
+def __dir__():
+    return __all__
+
+
+# See PEP 562 (https://peps.python.org/pep-0562/) for module-level __getattr__()
+def __getattr__(name):
+    if name in _SAMPLE_DATA:
+        return _get_sample_files([_SAMPLE_DATA[name]])[0]
+    if name == "file_dict":
+        return dict(
+            sorted(
+                zip(
+                    _SAMPLE_DATA.keys(),
+                    _get_sample_files(_SAMPLE_DATA.values(), no_download=True),
+                    strict=False,
+                )
+            )
+        )
+    if name == "file_list":
+        return [v for v in __getattr__("file_dict").values() if v]
+    msg = f"module '{__name__}' has no attribute '{name}'"
+    raise AttributeError(msg)
+
+
+def download_all(*, force_download=False):
+    """
+    Download all sample data at once that has not already been downloaded.
+
+    Parameters
+    ----------
+    force_download : `bool`
+        If ``True``, files are downloaded even if they already exist.  Default is
+        ``False``.
+    """
+    _get_sample_files(_SAMPLE_DATA.values(), force_download=force_download)