+++ aiapy/response/__init__.py
@@ -2,4 +2,4 @@
 Subpackage for AIA response functions.
 """
 
-from .channel import *  # NOQA
+from .channel import *  # NOQA: F403