+++ aiapy/util/__init__.py
@@ -2,5 +2,5 @@
 Subpackage with miscellaneous utility functions.
 """
 
-from .exceptions import *  # NOQA
-from .util import *  # NOQA
+from .exceptions import *  # NOQA: F403
+from .util import *  # NOQA: F403