+++ docs/conf.py
@@ -3,37 +3,13 @@ Configuration file for the Sphinx documentation builder.
 """
 
 import os
-import sys
-import datetime
-import warnings
-from pathlib import Path
 
-from packaging.version import Version
-
-# -- Read the Docs Specific Configuration --------------------------------------
 # This needs to be done before aiapy or sunpy is imported
 os.environ["PARFIVE_HIDE_PROGRESS"] = "True"
 
-# -- Check for dependencies ----------------------------------------------------
-from sunpy.util import missing_dependencies_by_extra  # NOQA: E402
-
-missing_requirements = missing_dependencies_by_extra("aiapy")["docs"]
-if missing_requirements:
-    print(  # NOQA: T201
-        f"The {' '.join(missing_requirements.keys())} package(s) could not be found and "
-        "is needed to build the documentation, please install the 'docs' requirements.",
-    )
-    sys.exit(1)
-
-# -- Project information -------------------------------------------------------
-project = "aiapy"
-author = "AIA Instrument Team"
-copyright = f"{datetime.datetime.now().year}, {author}"
-
-# Register remote data option with doctest
-import doctest  # NOQA: E402
-
-REMOTE_DATA = doctest.register_optionflag("REMOTE_DATA")
+import datetime  # NOQA: E402
+import warnings  # NOQA: E402
+from pathlib import Path  # NOQA: E402
 
 from astropy.utils.exceptions import AstropyDeprecationWarning  # NOQA: E402
 from matplotlib import MatplotlibDeprecationWarning  # NOQA: E402
@@ -42,17 +18,19 @@ from sunpy_sphinx_theme import PNG_ICON  # NOQA: E402
 
 from aiapy import __version__  # NOQA: E402
 
-# The full version, including alpha/beta/rc tags
+# -- Project information -------------------------------------------------------
+project = "aiapy"
+author = "AIA Instrument Team"
+copyright = f"{datetime.datetime.now(datetime.timezone.utc).year}, {author}"  # NOQA: A001
 release = __version__
-aiapy_version = Version(__version__)
-is_release = not (aiapy_version.is_prerelease or aiapy_version.is_devrelease)
+is_development = ".dev" in __version__
 
+# Need to make sure that our documentation does not raise any of these
 warnings.filterwarnings("error", category=SunpyDeprecationWarning)
 warnings.filterwarnings("error", category=SunpyPendingDeprecationWarning)
 warnings.filterwarnings("error", category=MatplotlibDeprecationWarning)
 warnings.filterwarnings("error", category=AstropyDeprecationWarning)
 
-# For the linkcheck
 linkcheck_ignore = [
     r"https://doi.org/\d+",
     r"https://element.io/\d+",
@@ -62,14 +40,6 @@ linkcheck_ignore = [
 linkcheck_anchors = False
 
 # -- General configuration -----------------------------------------------------
-# sphinxext-opengraph
-ogp_image = "https://raw.githubusercontent.com/sunpy/sunpy-logo/master/generated/sunpy_logo_word.png"
-ogp_use_first_image = True
-ogp_description_length = 160
-ogp_custom_meta_tags = [
-    '<meta property="og:ignore_canonical" content="true" />',
-]
-suppress_warnings = ["app.add_directive"]
 extensions = [
     "matplotlib.sphinxext.plot_directive",
     "sphinx_automodapi.automodapi",
@@ -101,6 +71,7 @@ default_role = "obj"
 napoleon_use_rtype = False
 napoleon_google_docstring = False
 napoleon_use_param = False
+suppress_warnings = ["app.add_directive"]
 nitpicky = True
 # This is not used. See docs/nitpick-exceptions file for the actual listing.
 nitpick_ignore = []
@@ -111,6 +82,50 @@ with Path("nitpick-exceptions").open() as nitpick_exceptions:
         dtype, target = line.split(None, 1)
         target = target.strip()
         nitpick_ignore.append((dtype, target))
+
+# -- Options for sphinxext-opengraph ------------------------------------------
+ogp_image = "https://raw.githubusercontent.com/sunpy/sunpy-logo/master/generated/sunpy_logo_word.png"
+ogp_use_first_image = True
+ogp_description_length = 160
+ogp_custom_meta_tags = [
+    '<meta property="og:ignore_canonical" content="true" />',
+]
+
+# -- Options for sphinx-copybutton ---------------------------------------------
+# Python Repl + continuation, Bash, ipython and qtconsole + continuation, jupyter-console + continuation
+copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
+copybutton_prompt_is_regexp = True
+
+# -- Options for hoverxref -----------------------------------------------------
+if os.environ.get("READTHEDOCS"):
+    hoverxref_api_host = "https://readthedocs.org"
+    if os.environ.get("PROXIED_API_ENDPOINT"):
+        # Use the proxied API endpoint
+        # - A RTD thing to avoid a CSRF block when docs are using a
+        #   custom domain
+        hoverxref_api_host = "/_"
+hoverxref_tooltip_maxwidth = 600  # RTD main window is 696px
+hoverxref_auto_ref = True
+hoverxref_mathjax = True
+# hoverxref has to be applied to these
+hoverxref_domains = ["py"]
+hoverxref_role_types = {
+    # roles with py domain
+    "attr": "tooltip",
+    "class": "tooltip",
+    "const": "tooltip",
+    "data": "tooltip",
+    "exc": "tooltip",
+    "func": "tooltip",
+    "meth": "tooltip",
+    "mod": "tooltip",
+    "obj": "tooltip",
+    # roles with std domain
+    "confval": "tooltip",
+    "hoverxref": "tooltip",
+    "ref": "tooltip",
+    "term": "tooltip",
+}
 
 # -- Options for intersphinx extension -----------------------------------------
 intersphinx_mapping = {
@@ -129,39 +144,6 @@ intersphinx_mapping = {
     ),
     "skimage": ("https://scikit-image.org/docs/stable/", None),
     "sunpy": ("https://docs.sunpy.org/en/stable/", None),
-}
-
-# -- Options for hoverxref -----------------------------------------------------
-if os.environ.get("READTHEDOCS"):
-    hoverxref_api_host = "https://readthedocs.org"
-
-    if os.environ.get("PROXIED_API_ENDPOINT"):
-        # Use the proxied API endpoint
-        # A RTD thing to avoid a CSRF block when docs are using a custom domain
-        hoverxref_api_host = "/_"
-
-hoverxref_auto_ref = False
-hoverxref_domains = ["py"]
-hoverxref_mathjax = True
-hoverxref_modal_hover_delay = 500
-hoverxref_tooltip_maxwidth = 600  # RTD main window is 696px
-hoverxref_intersphinx = list(intersphinx_mapping.keys())
-hoverxref_role_types = {
-    # Roles within the py domain
-    "attr": "tooltip",
-    "class": "tooltip",
-    "const": "tooltip",
-    "data": "tooltip",
-    "exc": "tooltip",
-    "func": "tooltip",
-    "meth": "tooltip",
-    "mod": "tooltip",
-    "obj": "tooltip",
-    # Roles within the std domain
-    "confval": "tooltip",
-    "hoverxref": "tooltip",
-    "ref": "tooltip",  # Would be used by hoverxref_auto_ref if we set it to True
-    "term": "tooltip",
 }
 
 # -- Options for HTML output ---------------------------------------------------
@@ -186,7 +168,6 @@ sphinx_gallery_conf = {
     "examples_dirs": Path("..") / "examples",
     "gallery_dirs": Path("generated") / "gallery",
     "matplotlib_animations": True,
-    # Comes from the theme.
     "default_thumb_file": PNG_ICON,
     "abort_on_example_error": False,
     "plot_gallery": "True",
@@ -194,8 +175,3 @@ sphinx_gallery_conf = {
     "doc_module": ("aiapy"),
     "only_warn_on_example_error": True,
 }
-
-# -- Options for sphinx-copybutton ---------------------------------------------
-# Python Repl + continuation, Bash, ipython and qtconsole + continuation, jupyter-console + continuation
-copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
-copybutton_prompt_is_regexp = True